+++ src/main/java/roomescape/dto/request/LogInRequest.java
@@ -1,7 +1,7 @@
 package roomescape.dto.request;
 
-import static roomescape.dto.InputValidator.validateNotBlank;
-import static roomescape.dto.InputValidator.validateNotNull;
+import static roomescape.dto.request.exception.InputValidator.validateNotBlank;
+import static roomescape.dto.request.exception.InputValidator.validateNotNull;
 
 public record LogInRequest(String email, String password) {
