+++ src/main/java/roomescape/dto/request/ThemeRequest.java
@@ -2,8 +2,8 @@ package roomescape.dto.request;
 
 import roomescape.domain.Theme;
 
-import static roomescape.dto.InputValidator.validateNotBlank;
-import static roomescape.dto.InputValidator.validateNotNull;
+import static roomescape.dto.request.exception.InputValidator.validateNotBlank;
+import static roomescape.dto.request.exception.InputValidator.validateNotNull;
 
 public record ThemeRequest(String name, String description, String thumbnail) {
