+++ src/test/java/roomescape/dto/InputValidatorTest.java
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.dto.request.exception.InputNotAllowedException;
+import roomescape.dto.request.exception.InputValidator;
 
 class InputValidatorTest {
